# Conflicts:
#	bridge/bridge/populate.py
#	bridge/bridge/vars.py
#	bridge/jobs/Download.py
#	bridge/jobs/models.py
#	bridge/jobs/urls.py
#	bridge/jobs/utils.py
#	bridge/marks/tables.py
#	bridge/marks/urls.py
#	bridge/reports/comparison.py
#	bridge/reports/utils.py

+++ bridge/marks/SafeUtils.py
@@ -61,6 +61,9 @@ class NewMark:
         if 'tags' not in self._args or not isinstance(self._args['tags'], list):
             raise ValueError('Unsupported tags: %s' % self._args.get('tags'))
 
+        if 'autoconfirm' in self._args and not isinstance(self._args['autoconfirm'], bool):
+            raise ValueError('Wrong type: autoconfirm (%s)' % type(self._args['autoconfirm']))
+
         tags = set(int(t) for t in self._args['tags'])
         if len(tags) > 0:
             tags_in_db = {}
@@ -97,9 +100,6 @@ class NewMark:
         return mark
 
     def change_mark(self, mark, recalculate_cache=True):
-        if len(self._args['comment']) == 0:
-            raise BridgeException(_('Change comment is required'))
-
         last_v = MarkSafeHistory.objects.get(mark=mark, version=F('mark__version'))
 
         mark.author = self._user
@@ -123,8 +123,9 @@ class NewMark:
         mark.save()
 
         if recalculate_cache:
-            MarkSafeReport.objects.filter(mark_id=mark.id).update(type=ASSOCIATION_TYPE[0][0])
-            SafeAssociationLike.objects.filter(association__mark=mark).delete()
+            if do_recalc or not self._args.get('autoconfirm', False):
+                MarkSafeReport.objects.filter(mark_id=mark.id).update(type=ASSOCIATION_TYPE[0][0])
+                SafeAssociationLike.objects.filter(association__mark=mark).delete()
             if do_recalc:
                 changes = ConnectMarks([mark]).changes
             else:
@@ -805,6 +806,14 @@ class PopulateMarks:
         presets_dir = os.path.join(settings.BASE_DIR, 'marks', 'presets', 'safes')
         new_marks = []
         for mark_settings in [os.path.join(presets_dir, x) for x in os.listdir(presets_dir)]:
+            identifier = os.path.splitext(os.path.basename(mark_settings))[0]
+            try:
+                MarkSafe.objects.get(identifier=identifier)
+                # The mark was already uploaded
+                continue
+            except ObjectDoesNotExist:
+                pass
+
             with open(mark_settings, encoding='utf8') as fp:
                 data = json.load(fp)
             if not isinstance(data, dict):
@@ -825,7 +834,7 @@ class PopulateMarks:
                 raise BridgeException(_('Corrupted preset safe mark: wrong description'))
             if not isinstance(data['is_modifiable'], bool):
                 raise BridgeException(_('Corrupted preset safe mark: is_modifiable must be bool'))
-            identifier = unique_id()
+
             new_marks.append(MarkSafe(
                 identifier=identifier, author=self._author, verdict=data['verdict'], status=data['status'],
                 is_modifiable=data['is_modifiable'], description=data['description'], type=MARK_TYPE[1][0]
