+++ bridge/marks/UnsafeUtils.py
@@ -75,6 +75,9 @@ class NewMark:
         if 'tags' not in self._args or not isinstance(self._args['tags'], list):
             raise ValueError('Unsupported tags: %s' % self._args.get('tags'))
 
+        if 'autoconfirm' in self._args and not isinstance(self._args['autoconfirm'], bool):
+            raise ValueError('Wrong type: autoconfirm (%s)' % type(self._args['autoconfirm']))
+
         tags = set(int(t) for t in self._args['tags'])
         if len(tags) > 0:
             tags_in_db = {}
@@ -110,8 +113,6 @@ class NewMark:
         return mark
 
     def change_mark(self, mark, recalculate_cache=True):
-        if len(self._args['comment']) == 0:
-            raise BridgeException(_('Change comment is required'))
         error_trace = None
         if 'error_trace' in self._args and isinstance(self._args['error_trace'], str):
             try:
@@ -156,8 +157,9 @@ class NewMark:
         mark.save()
 
         if recalculate_cache:
-            MarkUnsafeReport.objects.filter(mark_id=mark.id).update(type=ASSOCIATION_TYPE[0][0])
-            UnsafeAssociationLike.objects.filter(association__mark=mark).delete()
+            if do_recalc or not self._args.get('autoconfirm', False):
+                MarkUnsafeReport.objects.filter(mark_id=mark.id).update(type=ASSOCIATION_TYPE[0][0])
+                UnsafeAssociationLike.objects.filter(association__mark=mark).delete()
             if do_recalc:
                 changes = ConnectMarks([mark]).changes
             else:
@@ -788,6 +790,14 @@ class PopulateMarks:
 
         new_marks = []
         for mark_settings in [os.path.join(presets_dir, x) for x in os.listdir(presets_dir)]:
+            identifier = os.path.splitext(os.path.basename(mark_settings))[0]
+            try:
+                MarkUnsafe.objects.get(identifier=identifier)
+                # The mark was already uploaded
+                continue
+            except ObjectDoesNotExist:
+                pass
+
             with open(mark_settings, encoding='utf8') as fp:
                 data = json.load(fp)
             if not isinstance(data, dict):
@@ -814,7 +824,7 @@ class PopulateMarks:
                 raise BridgeException(_('Corrupted preset unsafe mark: comparison function name is required'))
             if data['comparison'] not in self._functions:
                 raise BridgeException(_('Preset unsafe mark comparison fucntion is not supported'))
-            identifier = unique_id()
+
             new_marks.append(MarkUnsafe(
                 identifier=identifier, author=self._author, verdict=data['verdict'], status=data['status'],
                 is_modifiable=data['is_modifiable'], description=data['description'], type=MARK_TYPE[1][0],
@@ -849,6 +859,7 @@ def delete_marks(marks):
         for report in changes[m_id]:
             unsafes_changes[report] = changes[m_id][report]
     RecalculateTags(unsafes_changes)
+    update_confirmed_cache(list(unsafes_changes))
     return unsafes_changes
 
 
@@ -895,7 +906,7 @@ def like_association(author, report_id, mark_id, dislike):
 
 
 def update_confirmed_cache(unsafes):
-    unsafes = list(safe.id for safe in unsafes)
+    unsafes = list(unsafe.id for unsafe in unsafes)
     with_confirmed = set(r_id for r_id, in MarkUnsafeReport.objects.filter(
         report_id__in=unsafes, type=ASSOCIATION_TYPE[1][0], error=None, result__gt=0).values_list('report_id'))
     ReportUnsafe.objects.filter(id__in=unsafes).update(has_confirmed=False)
